refactor: change `PasswordHasher` to `PasswordUtils`
